moved tube actions that were in run vuex, to tube vuex
