+++ src/components/PacbioLibraryEditModal.vue
@@ -63,7 +63,6 @@
 <script>
 
 import { createNamespacedHelpers } from 'vuex'
-// TODO: move to pacbio libraries
 const { mapActions } = createNamespacedHelpers('traction/pacbio/libraries')
 
 export default {
