+++ src/store/traction/pacbio/tubes/actions.js
@@ -44,14 +44,46 @@ const sampleExtractionTubeJson = (tubes) => {
   }))
 }
 
+const isLibraryBarcodeValid = async ({ dispatch }, barcode) => {
+  if (!barcode) { return false }
+  let libraryTube = await dispatch('getTubeForBarcode', barcode)
+  return validateLibraryTube(libraryTube)
+}
+
+const getTubeForBarcode = async ({ rootGetters }, barcode) => {
+  let request = rootGetters["traction/pacbio/tubes/tubeRequest"]
+  let promise = request.get({ filter: { barcode: barcode } })
+  let response = await handlePromise(promise)
+
+  if (response.successful && !response.empty) {
+    return response.deserialize.tubes[0]
+  }
+}
+
+const validateLibraryTube = (tube) => {
+  if (!tube) { return false }
+  if (!tube.materials) { return false }
+  if (!tube.materials.every(m => m.library_kit_barcode)) { return false }
+  // a way to validation libraries, as type is now container_material
+  // update test l.145
+  // if (tube.material.type != 'libraries') { return false }
+
+  return true
+}
+
 const actions = {
-  exportSampleExtractionTubesIntoTraction
+  exportSampleExtractionTubesIntoTraction,
+  isLibraryBarcodeValid,
+  getTubeForBarcode,
 }
 
 export {
   exportSampleExtractionTubesIntoTraction,
   sampleExtractionTubeJson,
-  processCostCode
+  processCostCode,
+  isLibraryBarcodeValid,
+  getTubeForBarcode,
+  validateLibraryTube,
 }
 
 export default actions
