+++ src/views/pacbio/PacbioLibraries.vue
@@ -115,7 +115,6 @@ import Alert from '@/components/Alert'
 import PrinterModal from '@/components/PrinterModal'
 import * as consts from '@/consts/consts'
 import { createNamespacedHelpers } from 'vuex'
-// TODO: move to pacbio libaries
 const { mapActions, mapGetters } = createNamespacedHelpers('traction/pacbio/libraries')
 
 export default {
